add junit for book service

+++ Library/src/main/java/com/semenihin/services/BookService.java
@@ -15,7 +15,7 @@ public interface BookService {
 
     Book findBook(long id);
 
-    void deleteBook(int id) throws LBFileAccessException;
+    void deleteBook(long id) throws LBFileAccessException;
 
     void printBooks();
